+++ src/app/components/columns-configuration-popover/column-editor.js
@@ -47,6 +47,12 @@ export default Component.extend(I18n, {
 
   /**
    * @virtual
+   * @type {PromiseObject<Array<{value: string, label: string}>>}
+   */
+  jsonKeyOptionsProxy: undefined,
+
+  /**
+   * @virtual
    * @type {Utils.ColumnsConfiguration}
    */
   columnsConfiguration: undefined,
@@ -73,7 +79,7 @@ export default Component.extend(I18n, {
    * @virtual optional
    * @type {string}
    */
-  initialJsonKey: '',
+  initialJsonKey: undefined,
 
   /**
    * @virtual optional
@@ -100,6 +106,8 @@ export default Component.extend(I18n, {
 
   xattrOptions: reads('xattrOptionsProxy.content'),
 
+  jsonKeyOptions: reads('jsonKeyOptionsProxy.content'),
+
   /**
    * @type {boolean}
    */
@@ -113,7 +121,7 @@ export default Component.extend(I18n, {
   /**
    * @type {string}
    */
-  newJsonKey: '',
+  jsonKeyFieldName: 'jsonKey',
 
   /**
    * @type {string}
@@ -135,7 +143,7 @@ export default Component.extend(I18n, {
     'initialDisplayedName',
     'metadataTypeValue',
     'jsonTypeField.value',
-    'newJsonKey',
+    'jsonKeyDropdownField.value',
     function newColumnName() {
       if (this.isColumnNameModified) {
         return this.columnName;
@@ -149,7 +157,7 @@ export default Component.extend(I18n, {
       if (this.jsonTypeField.value === 'all') {
         return defaultWholeJsonLabel;
       }
-      return this.newJsonKey;
+      return this.jsonKeyDropdownField.value;
     }
   ),
 
@@ -157,6 +165,7 @@ export default Component.extend(I18n, {
     'xattrKeyDropdownField',
     'metadataTypeField',
     'jsonTypeField',
+    'jsonKeyDropdownField',
     function fields() {
       return FormFieldsRootGroup
         .create({
@@ -168,6 +177,7 @@ export default Component.extend(I18n, {
             this.xattrKeyDropdownField,
             this.metadataTypeField,
             this.jsonTypeField,
+            this.jsonKeyDropdownField,
           ],
         });
     }
@@ -194,7 +204,6 @@ export default Component.extend(I18n, {
   metadataTypeField: computed(function metadataTypeField() {
     return RadioField
       .extend({
-        options: reads('columnEditorComponent.xattrOptions'),
         valueChanged() {
           this._super(...arguments);
           this.set('columnEditorComponent.isColumnNameModified', false);
@@ -234,6 +243,21 @@ export default Component.extend(I18n, {
       });
   }),
 
+  jsonKeyDropdownField: computed(
+    function jsonKeyDropdownField() {
+      return AutocompleteDropdownField
+        .extend({
+          options: reads('columnEditorComponent.jsonKeyOptions'),
+        })
+        .create({
+          columnEditorComponent: this,
+          name: this.jsonKeyFieldName,
+          defaultValue: this.initialJsonKey,
+          isOptional: true,
+        });
+    }
+  ),
+
   /**
    * @type {ComputedProperty<Utils.FormComponent.FormElement>}
    */
@@ -256,6 +280,13 @@ export default Component.extend(I18n, {
   }),
 
   /**
+   * @type {ComputedProperty<Utils.FormComponent.FormElement>}
+   */
+  jsonKeyDropdown: computed('fields', 'jsonKeyFieldName', function jsonKeyDropdown() {
+    return this.fields.getFieldByPath(this.jsonKeyFieldName);
+  }),
+
+  /**
    * @type {ComputedProperty<Boolean>}
    */
   isDisabledEditButton: bool('disabledEditButtonTooltip'),
@@ -270,13 +301,13 @@ export default Component.extend(I18n, {
     'xattrKeyDropdownField.value',
     'metadataTypeValue',
     'jsonTypeField.value',
-    'newJsonKey',
+    'jsonKeyDropdownField.value',
     function disabledEditButtonTooltip() {
       if (
         !this.newColumnName ||
         (!this.xattrKeyDropdownField.value && this.metadataTypeValue === 'xattr') ||
         (
-          !this.newJsonKey && this.metadataTypeValue === 'json' &&
+          !this.jsonKeyDropdownField.value && this.metadataTypeValue === 'json' &&
           this.jsonTypeField.value === 'key'
         )
       ) {
@@ -314,7 +345,7 @@ export default Component.extend(I18n, {
     'xattrKeyDropdownField.value',
     'jsonTypeField.value',
     'metadataTypeValue',
-    'newJsonKey',
+    'jsonKeyDropdownField.value',
     function isColumnAlreadyExisting() {
       let properties = {};
       if (this.metadataTypeValue === 'xattr') {
@@ -322,7 +353,7 @@ export default Component.extend(I18n, {
       } else {
         properties = {
           queryType: this.jsonTypeField.value,
-          jsonKey: this.newJsonKey,
+          jsonKey: this.jsonKeyDropdownField.value,
         };
       }
       return this.columnsConfiguration
@@ -337,9 +368,6 @@ export default Component.extend(I18n, {
   init() {
     initDestroyableCache(this);
     this._super(...arguments);
-    this.setProperties({
-      newJsonKey: this.initialJsonKey,
-    });
   },
 
   /**
@@ -367,7 +395,7 @@ export default Component.extend(I18n, {
       const type = this.metadataTypeValue;
       const options = type === 'xattr' ? { xattrKey: this.xattrKeyDropdownField.value } : {
         queryType: this.jsonTypeField.value,
-        jsonKey: this.newJsonKey,
+        jsonKey: this.jsonKeyDropdownField.value,
       };
       this.onSubmitColumn(
         this.isNewColumn,